macos: Generate universal binary in Makefile
This is currently done in package.sh, but this does not match how the
builds/signing is done: I want to do the builds in github, and then do
the signing. This works better if everything related to the build
including the universal binary generation is done in the Makefile.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>